First Commit.
This commit is copying the code I wrote for the class assignment, this is code written by me and was forced to be kept private while taking the class. Everything should be here and should be fine, however there might have to be other commits later to fix/change/add some things.